Supply Basic auth data
When a server offers Basic authentication, check for a user name
specified as an authentication option on the command-line, and for both
the user name and password in the environment (preferring the user name
value from the command-line if both are present), and if we don't have
both values, either prompt for or confirm the user name, and prompt for
the password to provide to the daemon.

Cache those answers so that we don't prompt more than once, even if a
given command-line invocation results in several connections and
requests to the daemon.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)