Call into libsasl to check Basic auth passwords
Use libsasl2 to check passwords if the "libsasl2=true" authn option is
used, offering it as an option for Basic authentication with a Basic
realm name which is either specified using the "realm" authn option, or
the default of "localhost".  The SASL service name used for the password
check is hard-coded to "docker".

Deployments which expect to use saslauthd for checking the passwords
will need to create a file named "docker.conf" under /usr/lib/sasl2 or
/etc/sasl2 containing this line:
  pwcheck_method:saslauthd

NB: When we move to building with a later version of the OS, we should
probably change the cgo directive that adds -lsasl2 to the LDFLAGS to
take advantage of pkg-config data, and update hack/make.sh to match.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)