Handle multi-step client-to-daemon Negotiate auth
Use the client's session to store an established-and-then-exported
GSSAPI context when we're authenticating clients using Negotiate.  This
requires working export/import functions in the implementation, but it
avoids having to maintain any of state on the server when accepting
authentication that requires multiple round trips.  It should also let
us recover the context later if we want to use the client's creds when
talking to a registry server.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)