Add a client-daemon authentication framework
Add the basic framework for requiring clients to authenticate to the
daemon when issuing requests to it.  This is mainly plumbing and wiring
in the logic in both the client and the server.

The client and server both contain module-local lists of functions which
create and return handlers for authentication schemes.  The
implementations register themselves at initialization time.  On clients,
the functions are used to create a set of AuthResponders which is
indexed in the client map by the authentication schemes that they
implement.  On servers, the functions are used to create a list of
Authenticators in each server which has authentication enabled, each of
which gets a crack at every client request.

The client will call the AuthResponder for a particular scheme if it
receives a 401 error, and if it signals that the client should try
again, it will do so, repeating the process until it receives a
successful result.  At that point, it will call the scheme's
AuthCompleted callback to check if the server's reply is acceptable to
it.

The server will call either the GetChallenge methods or CheckResponse
methods of its Authenticators for each request, depending on whether or
not the incoming request includes any Authorization headers.  If
CheckResponse indicates that the client's identity has been verified,
information about the authenticated user, and the name of the
authentication scheme that succeeded, will be added to the request
structure as a shared value so that it can be used later when we call
into authorization plugins.  The middleware for authorization is moved
earlier in order to get the ordering right at run time.

The server's RequireAuthn flag can be found in its configuration
structure, and both client and server get an opportunity to retrieve
option values, which are passed to the --authn-opt flag on the command
line, from a map named AuthnOpts in the server configuration structure,
or a map named authnOpts in the DockerCli object.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)

Fixed a bug which disabled persistent connections when docker authentication is required

Signed-off-by: lior <lior@twistlock.com>