Add authentication via external plugins
Add a call to the Authentication.GetChallenge() API provided by one or
more "Authentication" plugins, whose names can be passed to the daemon
as a list of values for the "plugins" authentication option, and which
provide headers to send back to a client that sent us a request with no
"Authorization" header.  Call the plugin Authentication.CheckResponse()
methods when the client request includes at least one "Authorization"
header, and if any plugin returns a usable User structure, accept it as
proof of the client's identity.

We can't do the client part of authentication using plugins, because
it's not really useful to require the client to use plugins which have
to be registered in a central location and ensure that a plugin is
running, so the client still has to be taught how to respond to
particular authentication requirements directly.

Some of the server bits we add later in the patch series could be moved
out to plugins.  We're already linking with libraries for the client's
sake, so using a plugin for Negotiate instead of calling the APIs
directly doesn't avoid the linking dependency.  The client socket can't
be passed to a plugin, so looking up the user's identity using the
socket descriptor and some help from the kernel is not something that
can be done via a plugin.

That said, plugins remain a viable method for implementing additional
ways to check passwords for Basic authentiation, for verifying Bearer
tokens, and for supporting other schemes which the connecting client may
support.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)