Log directly to journald
Instead of depending on our stdout/stderr descriptors, which are
connected to a running journald at startup, start logging messages
generated by the daemon directly to journald as soon as we can.

This lets us continue to log when journald is restarted, and it lets us
turn off the redundant timestamp for just those cases.  We cheat a bit
by triggering based on whether or not $NOTIFY_SOCKET is set in the
environment, but it avoids having to set a custom value in the unit file
to control the behavior.

Signed-off-by: Nalin Dahyabhai (nalin@redhat.com) (github: nalind)