Teach the daemon to use plugins for authentication
Add a plugin-based authentication middleware to the server.

Logic which pulls the common name out of the peer's certificate when TLS
is in use is moved from the authorization middleware to the
authentication package so that we can avoid attempting to authenticate
the client at the HTTP level in cases where we're just going to end up
using the certificate anyway.

The middleware handles everything using the configured set of plugins,
which are wrapped up in an object that exposes methods for gathering the
set of challenges to send to clients, and for evaluating a client's
response.

The server will call the Authentication method of each plugin for each
request, passing it the full set of headers from the request, along with
selected information from the Request object, and the method can return
headers which should be added to the server's response and/or
information about a successfully authenticated client.

The server's RequireAuthentication flag can be found in its
configuration structure, and both client and server get an opportunity
to retrieve option values, which are passed to the --authn-opt flag on
the command line, from a map named AuthenticationOoptions in the server
configuration structure.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)

Fixed a bug which disabled persistent connections when docker
authentication is required

Signed-off-by: lior <lior@twistlock.com>