Remember defaults in the devmapper graph driver
When we successfully initialize the devmapper graph driver, collect the
options that we used and save them to a file named "defaults".  Prepend
that file's contents, if it exists and we can read it, to the list of
options that we're passed when initializing the devmapper graph driver,
returning an error only for unrecognized options that we didn't recover
from the file.